CleanUp and respect rule IDE0034 (simplified default expression)

+++ Microsoft.CodeAnalysis/CodeGen/BasicBlock.cs
@@ -9,9 +9,8 @@ using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Metadata;
-using Microsoft.CodeAnalysis.Debugging;
+
 using Microsoft.CodeAnalysis.PooledObjects;
-using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CodeGen
 {
@@ -637,7 +636,7 @@ namespace Microsoft.CodeAnalysis.CodeGen
                             break;
                     }
 
-                    return (int)RegularInstructionsLength + branchSize;
+                    return RegularInstructionsLength + branchSize;
                 }
             }
