+++ Microsoft.CodeAnalysis/Collections/SegmentedList_T1.cs
@@ -14,6 +14,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+
 using Microsoft.CodeAnalysis.Collections.Internal;
 
 namespace Microsoft.CodeAnalysis.Collections
@@ -1091,7 +1092,7 @@ namespace Microsoft.CodeAnalysis.Collections
         //
         public void TrimExcess()
         {
-            var threshold = (int)(((double)_items.Length) * 0.9);
+            var threshold = (int)(_items.Length * 0.9);
             if (_size < threshold)
             {
                 Capacity = _size;
